requirements.txt

fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.5.0
websockets>=12.0


langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.0.1
langchain-openai>=0.0.5
langchainhub>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langgraph>=0.0.15


openai>=1.0.0
sentence-transformers>=2.2.2


chromadb>=0.4.22


sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0
alembic>=1.12.0
aiomysql>=0.2.0
asyncmy>=0.2.10


aiosqlite>=0.20.0
motor>=3.3.0


pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.2


pypdf>=3.17.0
python-docx>=1.1.0
docx2txt>=0.8
beautifulsoup4>=4.12.0


httpx>=0.27.0
requests>=2.31.0


playwright>=1.45.0
aiofiles>=23.2.1
typer>=0.12.3
Pillow>=10.0.0
tenacity>=8.2.0
opencv-python>=4.8.0
redis>=4.6.0
jieba>=0.42.1
wordcloud>=1.9.0
matplotlib>=3.7.0
parsel>=1.9.0
pyexecjs>=1.5.1
pyhumps>=3.8.0


python-dotenv>=1.0.0
tqdm>=4.66.0


streamlit>=1.29.0
plotly>=5.18.0
streamlit-aggrid>=0.3.4


jupyter>=1.0.0

ipython>=8.0.0


mcp>=1.0.0
